refactor: change loading component for form-exercise-page and changed file names
