+++ src/features/module/components/exercise-page-skeleton.tsx
@@ -1,0 +1,15 @@
+export default function FormExercisePageSkeleton() {
+  return (
+    <div className="p-2 lg:px-80">
+      <div className="mx-12 mt-18 h-12 animate-pulse rounded-xl bg-zinc-200 md:mx-auto md:w-120 lg:mx-auto lg:mt-16 lg:w-145" />
+
+      <div className="m-2 mt-20 h-12 animate-pulse rounded-md bg-zinc-200 lg:mt-22" />
+      <div className="m-2 mt-15 h-12 animate-pulse rounded-md bg-zinc-200 lg:mt-14" />
+      <div className="m-2 mt-15 h-12 animate-pulse rounded-md bg-zinc-200" />
+
+      <div className="m-2 mt-15 h-40 animate-pulse rounded-md bg-zinc-200" />
+
+      <div className="m-4 mt-16 mb-60 h-16 animate-pulse rounded-md bg-zinc-200" />
+    </div>
+  );
+}
