+++ src/features/module/components/introduction-page-skeleton.tsx
