+++ src/features/module/pages/form-exercise-page.tsx
@@ -7,6 +7,7 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { useEffect } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
 import { useParams } from 'react-router';
+import FormExercisePageSkeleton from '../components/exercise-page-skeleton';
 import { FormSectionWrapper } from '../components/form-section-wrapper';
 import { ModuleHeader } from '../components/module-header';
 import { useCreateTranslateExercise } from '../hooks/api/mutations/use-create-translate-exercise';
@@ -130,9 +131,8 @@ export const FormExercisePage = () => {
     }
   };
 
-  //TODO: Fazer componente de loading
   if (isLoading) {
-    return <div>carregando</div>;
+    return <FormExercisePageSkeleton />;
   }
 
   return (
