+++ src/features/module/pages/form-module-page.tsx
@@ -8,9 +8,9 @@ import { useNavigate, useParams } from 'react-router';
 import { Back } from '../../../components/back';
 import { BannerUpload } from '../components/banner-upload';
 import { DescriptionField } from '../components/description-field';
-import FormModulePageSkeleton from '../components/form-module-page-skeleton';
 import { FormSectionWrapper } from '../components/form-section-wrapper';
 import { GrammarRulesField } from '../components/grammar-rules-field';
+import FormModulePageSkeleton from '../components/module-page-skeleton';
 import { TitleInput } from '../components/title-input';
 import { useCreateModule } from '../hooks/api/mutations/use-create-module';
 import { useUpdateModule } from '../hooks/api/mutations/use-update-details';
