+++ src/features/module/pages/form-presentation-page.tsx
@@ -8,9 +8,9 @@ import { Info } from 'lucide-react';
 import { useEffect } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
 import { useParams } from 'react-router';
-import FormPresentationPageSkeleton from '../components/form-presentation-page-skeleton';
 import { FormSectionWrapper } from '../components/form-section-wrapper';
 import { ModuleHeader } from '../components/module-header';
+import FormPresentationPageSkeleton from '../components/presentation-page-skeleton';
 import { useCreatePresentation } from '../hooks/api/mutations/use-create-presentation';
 import { useUpdatePresentation } from '../hooks/api/mutations/use-update-presentation';
 import { useGetPresentation } from '../hooks/api/queries/use-get-presentation';
